Added utility for HtLodash which combines the previous hidash util and decorates necessary sprintf and underscore.string methods onto lodash.
