+++ lib/server/middleware/parse-tag-hash.js
@@ -6,7 +6,7 @@
  * @param _
  * @returns {tagMiddleware}
  */
-var _ = require('lodash');
+
 module.exports = function(wilsonConfig, ConnectionFilterUtil, logger) {
 
   function tagMiddleware(req, res, next) {
