+++ lib/server/middleware/tags.js
@@ -4,9 +4,9 @@
  * @returns {tagMiddleware}
  */
 
-var path = require('path'),
-  fs = require('fs'),
-  _ = require('lodash'),
+var path    = require('path'),
+  fs        = require('fs'),
+  _         = require('../utils/HtLodash'),
   walkStack = require('../utils/walk-stack');
 
 module.exports = function(wilsonConfig) {
