+++ lib/server/middleware/wilson-locals.js
@@ -5,15 +5,15 @@
  * @param res
  * @param next
  */
-var _ = require('lodash');
+var _ = require('../utils/HtLodash');
 
-module.exports = function(wilsonConfig, hidash) {
+module.exports = function(wilsonConfig) {
   function configLocalsMiddlware(req, res, next) {
     var config = _.cloneDeep(wilsonConfig);
     //console.log('config', config);
 
     //merge req config info to the client config
-    hidash.mergeAll(config.client, req.wilson);
+    _.mergeAll(config.client, req.wilson);
 
     res.locals.wilsonConfig = config;
