+++ lib/server/modules/I18NModule.js
@@ -1,14 +1,14 @@
-var i18next = require('i18next-node-stable'),
-    asyncUtil = require('async'),
-    _ = require('lodash'),
-    fs = require('fs'),
-    hbs = require('hbs'),
-    Q = require('q'),
-    strip    = require('strip-json-comments'),
-    url = require('url'),
-    path = require('path');
+var i18next       = require('i18next-node-stable'),
+    asyncUtil     = require('async'),
+    _             = require('../utils/HtLodash'),
+    fs            = require('fs'),
+    hbs           = require('hbs'),
+    Q             = require('q'),
+    strip         = require('strip-json-comments'),
+    url           = require('url'),
+    path          = require('path');
 
-module.exports = function(wilsonConfig, logger, ComponentService, hidash) {
+module.exports = function(wilsonConfig, logger, ComponentService) {
   // any path that starts with xx or xx-YY is considered a locale
   var PATH_LOCALE_REGEXP = /^([a-z]{2}|[a-z]{2}-[a-zA-Z]{2})$/i;
 
@@ -47,8 +47,8 @@ module.exports = function(wilsonConfig, logger, ComponentService, hidash) {
     debug: false
   };
 
-  //Merge in Wilson config settings with defaults
-  hidash.mergeAll(i18nextOptions, wilsonConfig.client.i18n);
+  // Merge in Wilson config settings with defaults
+  _.mergeAll(i18nextOptions, wilsonConfig.client.i18n);
 
 
   /**
@@ -173,7 +173,7 @@ module.exports = function(wilsonConfig, logger, ComponentService, hidash) {
 
             //Merge the translations on top of any previous ones
             if (resStore[lng] && resStore[lng][namespace]) {
-              hidash.mergeAll(resStore[lng][namespace], namespaceJson);
+              _.mergeAll(resStore[lng][namespace], namespaceJson);
             } else {
               resStore[lng][namespace] = namespaceJson;
             }
