+++ lib/server/routes/component.js
@@ -1,4 +1,3 @@
-var _ = require('lodash');
 
 module.exports = function(wilsonConfig, ComponentService, ConnectionFilterUtil, logger) {
   /**
