+++ lib/server/routes/config.js
@@ -2,10 +2,10 @@
  * ENDPOINT: /config
  */
 
-var path = require('path'),
-    fs = require('fs'),
-    _ = require('lodash'),
-    hbs = require('hbs');
+var path  = require('path'),
+    fs    = require('fs'),
+    _     = require('../utils/HtLodash'),
+    hbs   = require('hbs');
 
 module.exports = function(wilsonConfig, logger) {
   //Load index page
