+++ lib/server/services/BundleService.js
@@ -6,7 +6,7 @@
  */
 var fs    = require("fs"),
     path  = require("path"),
-    _     = require('lodash');
+    _     = require('../utils/HtLodash');
 
 module.exports = function(wilsonConfig, logger) {
   //console.log('logger(BundleService)', logger);
