+++ lib/server/services/ComponentService.js
@@ -24,13 +24,9 @@ var fs            = require("fs"),
     ProcessUtil   = require('../utils/ProcessUtil'),
     ArrayUtil     = require('../utils/ArrayUtil'),
     ustring       = require('underscore.string'),
-    _             = require('lodash');
+    _             = require('../utils/HtLodash');
 
-// Add sprintf function to lodash
-_.sprintf         = require('sprintf-js').sprintf;
-_.strRight        = ustring.strRight;
-
-module.exports = function(wilsonConfig, wilsonFrameworkConfig, logger, CacheService, BundleService, RevisionedFileUtil, hidash) {
+module.exports = function(wilsonConfig, wilsonFrameworkConfig, logger, CacheService, BundleService, RevisionedFileUtil) {
   //console.log('logger(ComponentService)', logger);
   //console.log('CacheService', CacheService);
