+++ lib/server/utils/ArrayUtil.js
@@ -1,4 +1,4 @@
-var _ = require('lodash');
+var _ = require('./HtLodash');
 
 var concatNonNull = function() {
   var newArray = [];
