+++ lib/server/utils/ConnectionFilterUtil.js
@@ -7,12 +7,7 @@
  * @copyright (c) 2013 Hightail Inc. All Rights Reserved
  */
 
-var _     = require('lodash'),
-   //UserAgentUtil = require('./UserAgentUtil'),
-   //UrlUtil = require('../util/UrlUtil'),
-   //uaParser      = require('ua-parser'),
-   //RequestParser = require('./RequestParser'),
-   //uuid           = require('node-uuid'),
+var _       = require('./HtLodash'),
    LZString = require('lz-string');
