+++ lib/server/utils/RevisionedFileUtil.js
@@ -1,4 +1,4 @@
-var _ = require('lodash'),
+var _ = require('./HtLodash'),
   path = require('path'),
   qfs = require('q-io/fs'),
   Q = require('q'),
