+++ lib/server/utils/UrlUtil.js
@@ -1,9 +1,4 @@
-var _       = require('lodash'),
-    ustring = require('underscore.string');
-
-
-// Decorate required functions
-_.strRightBack = ustring.strRightBack;
+var _ = require('./HtLodash');
 
 var PATH_CHARS = ' /';
