+++ lib/server/wilson.js
@@ -1,16 +1,15 @@
-var express = require('express');
+var express               = require('express');
 
 //DI
-var dependable = require('dependable'),
-  container = dependable.container();
+var dependable            = require('dependable'),
+  container               = dependable.container();
 
-var path = require('path'),
-    hbs = require('hbs'),
-    fs = require('fs'),
-    _ = require('lodash'),
-    hidash = require('./utils/hidash'),
-    noCacheMiddleware = require('./middleware/no-cache'),
-    walkStack = require('./utils/walk-stack');
+var path                  = require('path'),
+    hbs                   = require('hbs'),
+    fs                    = require('fs'),
+    _                     = require('./utils/HtLodash'),
+    noCacheMiddleware     = require('./middleware/no-cache'),
+    walkStack             = require('./utils/walk-stack');
 
 //Load in the default Wilson config
 var DEFAULT_WILSON_CONFIG = require(path.join(__dirname, './wilson-config.json'));
@@ -22,7 +21,7 @@ var DEFAULT_WILSON_CONFIG = require(path.join(__dirname, './wilson-config.json')
  */
 function prepareConfig(appConfig) {
   //Merge the passed in config options ON TOP of the defaults
-  var config = hidash.mergeAll({}, DEFAULT_WILSON_CONFIG, appConfig);
+  var config = _.mergeAll({}, DEFAULT_WILSON_CONFIG, appConfig);
 
   //Resolve the root path so that it can be used easily from other modules
   config.server.projectPaths.root = path.resolve(config.server.projectPaths.root);
@@ -64,7 +63,6 @@ module.exports = function(app, appWilsonConfigJson, dependencies) {
   container.register('I18NModule', require('./modules/I18NModule'));
 
   //Utilities
-  container.register('hidash', hidash);
   container.register('ConnectionFilterUtil', require('./utils/ConnectionFilterUtil'));
   container.register('RevisionedFileUtil', require('./utils/RevisionedFileUtil'));
